CLOUD-548 removed service logic from the stackconverter. Added stack decorator
