+++ src/main/resources/logback.xml
@@ -66,7 +66,7 @@
     <logger name="org.hibernate.type.descriptor.sql.BasicBinder">
         <level value="OWNER_BASED"/>
     </logger>
-    <root level="DEBUG">
+    <root level="INFO">
         <appender-ref ref="OWNER_BASED"/>
     </root>
